requirements.txt
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-dotenv==1.0.0
openai==1.12.0
PyPDF2==3.0.1
